Current version of Incident Response Workshop Materials
This includes:
 - Slides for the workshop
 - CC BY 2.0 photos used in the slides
 - Code of Conduct for the workshop
 - Pre-work for creating a safe(r) event
 - Incident report template for responders
 - Volunteer incident guide

This does not include the roleplay prompts for the workshop. In order to
ensure attendees can respond "in the moment" during the role-play
practice, this document is withheld from attendees until after they
complete the workshop. Attendees will be given access to a private
repository with the current version of the role-play documents.