Use new QUEST acronym for reporter template
I've gotten feedback from several workshop attendees that I need to go
over the report template before they practice taking a report. In
putting that into the slides, I realized the LASER acronym is not really
working well.

The first parts, listen and acknowledge emotions were fine. I wanted to
talk about getting help in an unsafe situation sooner. Ensuring people
are safe is important for the conversation with the reporter. The
"report" or "respond" bit didn't really work, since that's skipping
ahead in the material. I also had a lot of important material around the
acronym that didn't fit, like taking the person to a quiet place,
evaluating reporter privacy, etc.

The new acronym is "QUEST", and it encompasses the information that
didn't fit into the previous acronym, and drops the unimportant bits.
Listening is still very important, but that's something I think people
will understand?

I'll try this out at my next workshop, and see how it goes.

Additionally, I modified the privacy slide to include the option of
telling the reported person that a witness overheard the conversation or
saw the incident, rather than the person involved who actually reported
it.